+++ src/lib.rs
@@ -319,7 +319,7 @@ captured.
 
 ## Builtin Structure Matchers
 
-A bit different but similar to the structure selectors `$S ...` are the `$P
+A bit different but similar to the structure selectors `$S ...` are the `$M
 ...` or `match` structure matchers:
 
 ```wlambda
@@ -358,7 +358,7 @@ std:assert_str_eq res $[
 !crate  = $none;
 !domain = $none;
 
-? some_url <& $r{$^ (^$+[^:]) \:\/\/ (^$*[^/]) \/crates\/ (^$+[a-z]) } {
+? some_url &> $r{$^ (^$+[^:]) \:\/\/ (^$*[^/]) \/crates\/ (^$+[a-z]) } {
     .domain = $\.2;
     .crate = $\.3;
 };